+++ update_manager/install_saturn_go_nginx.sh
@@ -238,6 +238,7 @@ copy_web_asset "saturngo.html"
 copy_web_asset "p23test.html"
 copy_web_asset "fpga.html"
 copy_web_asset "pihpsdr.html"
+copy_web_asset "deskhpsdr.html"
 
 if [[ -f "$SOURCE_DIR/scripts/config.json" ]]; then
   cp -f "$SOURCE_DIR/scripts/config.json" "$WEB_ROOT/config.json"